Fixed crashes, Improved obtaining of radius of region.
Fixed issue of crashing caused due to negative sign for force. This
caused the atoms to move outside the region. Estimation of radius is now
more robust- Earlier the radius was calculated using simulation box
dimensions, this means that if the circular cross section of the region
is smaller than the box cross section then an incorrect radius would be
used. To fix this the dimensions of the region is used directly. Removed
the hypergeometric function in favour of using the hypergeometric_2F1
function as the latter has very good agreement with Wolfram Mathematica
unlike the former for certain cases. Improved documentation.